Create Phylogeny
The newly formed tree is less similar to the fungi reference tree.

A potential reason: The structure of phylogenetic tree depends on the orthologous species and how closely they are related to each other. The less diverged the species are, the more closely related the species will be.

In the new tree, the species included could be more diverged from each other and are not orthologs anymore, meaning the orthology between the species might have been lost at some point of branching out. The branches of the new phylogenetic tree doesn't recapitulate the sequence of speciation events.

Another potential reason: Reciprocal Best Match (RBM) finds functionally most similar or orthologs.

In the new tree, RBM could be low, that's why it couldn't find more orthologous genes.

Another potential reason: Duplication events.

During branching in the new phylogenetic tree, it is possible that few or more duplication events took place, that's why it is less similar to the reference fungi tree.